umip: use 64-bit variables for expected results
Some of the tests involve 64-bit operands. Hence, we should use
64-bit variables to store the expected results in order to perform
results' validation correctly.

Also, update printf modifiers accordingly.

+++ meta-luv/recipes-core/umip/files/umip_test_defs.h
@@ -104,9 +104,9 @@ struct table_desc {
 	unsigned long base;
 } __attribute__((packed));
 
-static const unsigned short expected_msw = EXPECTED_SMSW;
-static const unsigned short expected_ldt = EXPECTED_SLDT;
-static const unsigned short expected_tr = EXPECTED_STR;
+static const unsigned long expected_msw = EXPECTED_SMSW;
+static const unsigned long expected_ldt = EXPECTED_SLDT;
+static const unsigned long expected_tr = EXPECTED_STR;
 
 static const struct table_desc expected_gdt = {
 	.limit = EXPECTED_GDT_LIMIT,
